Yeah
Subtopics looked like amazing. Like, the best they’ve ever looked.  But
in the process of doing that I screwed up the styling for like
everything that wasn’t a Subtopic. So I had to go back. I spent like 30
mins on it, it was amazing.  This commit isn’t really necessary; but I
feel like future generations need to hear this story. So if you’re
wondering why I committed my code to change a cursor to “pointer”, that
is why. I am so mad at myself rn.  I’m probably gonna make it all cool
again later, when it is not 2 Am. Goodnight.